feat: add functions for hash calculation
